Int type images

+++ tests/utils/test_saver.py
@@ -6,7 +6,7 @@ from tf_explain.utils.saver import save_grayscale, save_rgb
 
 
 def test_should_save_grayscale_image(output_dir):
-    input_image = np.ones((28, 28, 1))
+    input_image = np.ones((28, 28, 1), dtype="uint8")
 
     save_grayscale(input_image, output_dir, "grayscale.png")
